feat: refactor hook useAlertBar and add close icon
